IMPLEMENTED THE SCREEN CONTROLLER FUNCTION
The screen controller function is what makes users able to interact with the game from the DOM instead of the console.
Created an updateScreen function that prints the gameboard on the DOM and and adds an eventlistener clickHandlerBtn to each cell that help changes the value of that cell to the value associated with the player token
Created the reset button, so we can start a new game.